requirements.txt
django
django-allauth
asgiref
celery
certifi
cffi
channels
channels-redis
charset-normalizer
cryptography
idna
psycopg2-binary
pycparser
pyjwt
python-dotenv
redis
requests
sqlparse
urllib3
aiohttp
pytest